requirements.txt
astroid==2.11.7
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.5.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pathlib==1.0.1
Pillow==9.2.0
platformdirs==2.5.2
pylint==2.14.5
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
